fix: foi refeito para poder passar informações separadamentes

+++ src/services/InfoUserServices.ts
@@ -2,10 +2,10 @@ import { prisma } from "../config/prisma";
 
 interface InfoUserProps {
   userId: number;
-  bio: string;
-  regiao: string;
-  profissao: string;
-  skills: string[];
+  bio?: string;
+  regiao?: string;
+  profissao?: string;
+  skills?: string[];
 }
 
 export const InfoUserServices = async ({
@@ -15,26 +15,34 @@ export const InfoUserServices = async ({
   profissao,
   skills,
 }: InfoUserProps) => {
+  const data: Record<string, any> = {};
+
+  if (bio !== undefined) {
+    data.bio = bio;
+  }
+
+  if (regiao !== undefined) {
+    data.regiao = regiao;
+  }
+
+  if (profissao !== undefined) {
+    data.profissao = profissao;
+  }
+
+  if (skills !== undefined) {
+    data.skills = {
+      deleteMany: {},
+      create: skills.map((skill) => ({
+        name: skill,
+      })),
+    };
+  }
 
   return prisma.user.update({
     where: {
       id: userId,
     },
-
-    data: {
-      bio,
-      regiao,
-      profissao,
-
-      skills: {
-        deleteMany: {},
-
-        create: skills.map((skill) => ({
-          name: skill,
-        })),
-      },
-    },
-
+    data,
     include: {
       skills: true,
       services: true,
